pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fileforge"
version = "1.0.0"
description = "Intelligent file processing and analysis platform with LLM integration"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "FileForge Team", email = "team@fileforge.dev"}
]
keywords = ["file-processing", "llm", "ocr", "vision", "pipeline"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Text Processing :: General",
]

dependencies = [
    # CLI framework
    "typer>=0.12.0",
    "rich>=13.0.0",

    # Configuration management
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tomli-w>=1.0.0",

    # Plugin system
    "pluggy>=1.5.0",

    # File watching
    "watchdog>=6.0.0",

    # Windows support
    "pywin32>=311; sys_platform == 'win32'",

    # LLM integration
    "ollama>=0.3.0",

    # Vector database
    "chromadb>=0.4.0",

    # SQLite vector search
    "sqlite-vec>=0.1.0",

    # Document processing
    "pymupdf>=1.26.0",
    "python-docx>=1.1.0",

    # OCR engines
    "paddleocr>=2.7.0",
    "pytesseract>=0.3.10",

    # Computer vision
    "ultralytics>=8.0.0",

    # Face detection
    "deepface>=0.0.79",

    # NSFW detection
    "nudenet>=3.0.0",

    # Common utilities
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-pillow",
]

[project.scripts]
fileforge = "fileforge.cli:app"

[project.urls]
Homepage = "https://github.com/fileforge/fileforge"
Documentation = "https://docs.fileforge.dev"
Repository = "https://github.com/fileforge/fileforge"
"Bug Tracker" = "https://github.com/fileforge/fileforge/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["fileforge*"]
namespaces = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=fileforge",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "paddleocr.*",
    "pytesseract.*",
    "ultralytics.*",
    "deepface.*",
    "nudenet.*",
    "chromadb.*",
    "ollama.*",
    "pluggy.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/fileforge"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
